@@ -14,16 +14,16 @@
 	<parent>
 		<groupId>org.springframework.boot</groupId>
 		<artifactId>spring-boot-starter-parent</artifactId>
-		<version>2.1.2.RELEASE</version>
+		<version>3.4.3</version>
 	</parent>
 
 	<properties>
 		<project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
 		<project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-		<java.version>1.8</java.version>
+		<java.version>17</java.version>
 		<lombok.version>1.16.10</lombok.version>
         <okhttp.version>2.7.5</okhttp.version>
-        <spring.boot.version>2.1.2.RELEASE</spring.boot.version>
+        <spring.boot.version>3.4.3</spring.boot.version>
 	</properties>
 
 	<dependencies>
@@ -39,10 +39,10 @@
 		</dependency>
 
 		<dependency>
-    		<groupId>org.mockito</groupId>
-    		<artifactId>mockito-core</artifactId>
-    		<version>2.13.0</version>
-    		<scope>test</scope>
+			<groupId>org.mockito</groupId>
+			<artifactId>mockito-core</artifactId>
+			<version>5.12.0</version>
+			<scope>test</scope>
 		</dependency>
 
 		<dependency>
@@ -55,46 +55,40 @@
 		    <artifactId>spring-boot-starter-actuator</artifactId>
 		  </dependency>
 
-		<dependency>
-		    <groupId>io.springfox</groupId>
-		    <artifactId>springfox-swagger2</artifactId>
-		    <version>2.7.0</version>
-		</dependency>
-
-		<dependency>
-		    <groupId>io.springfox</groupId>
-		    <artifactId>springfox-swagger-ui</artifactId>
-		    <version>2.10.0</version>
-		</dependency>
+		 <dependency>
+            <groupId>org.springdoc</groupId>
+            <artifactId>springdoc-openapi-starter-webmvc-ui</artifactId>
+            <version>2.1.0</version>
+        </dependency>
 
 		<dependency>
 			<groupId>com.microsoft.azure</groupId>
 			<artifactId>azure-servicebus</artifactId>
-			<version>1.2.8</version>
+			<version>3.6.7</version>
 		</dependency>
 
         <dependency>
             <groupId>com.microsoft.azure</groupId>
             <artifactId>applicationinsights-spring-boot-starter</artifactId>
-            <version>1.1.1</version>
+            <version>2.6.2</version>
         </dependency>
 
         <dependency>
             <groupId>com.microsoft.azure</groupId>
             <artifactId>applicationinsights-web</artifactId>
-            <version>2.3.0</version>
+            <version>2.6.2</version>
         </dependency>
 
 		<dependency>
 			<groupId>com.microsoft.azure</groupId>
 			<artifactId>applicationinsights-logging-logback</artifactId>
-			<version>2.3.0</version>
+			<version>2.6.2</version>
 		</dependency>
 
         <dependency>
             <groupId>org.bouncycastle</groupId>
-            <artifactId>bcprov-jdk15on</artifactId>
-            <version>1.67</version>
+            <artifactId>bcprov-jdk18on</artifactId>
+            <version>1.80</version>
         </dependency>
     </dependencies>
 
